agent/unix: add dependency between ovs interface and port
When a new port is added to the bridge agent creates an interface to
this port. Since CS does not synchronize after the add command
this instance does not get into the CS database. In order to fix it
interface node must be dependent on /agent/ovs/bridge/port node.

Signed-off-by: Artemii Morozov <artemii.morozov@arknetworks.am>